requirements.txt
python-dateutil==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
python-telegram-bot[webhooks]==21.1.1
pytz==2024.1